Changed overlay layer order.
Moved map related utils to maputils module.
