+++ src/app/shared/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './consts';
+export * from './maputils';
